requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
authlib
python-jose
passlib
itsdangerous
httpx
